Python file
The given file contains the basic codes of python.
Here, firstly the libraries are imported which are required. Then Salaries csv file is uploaded and some basic functions are performed to know the count of elements in file, to know the column names in the file and also their type and first few rows were called in the file through df.head function.